cleaned up and organized codebase
