+++ package.json
@@ -33,7 +33,7 @@
     "claude",
     "copilot"
   ],
-  "author": "Janardhan Pollle <www.janardhanpolle@gmail.com>",
+  "author": "Janardhan Pollle",
   "license": "MIT",
   "repository": {
     "type": "git",
